Replace rotated cubes with hexagonal cones
The rotated cubes ended up with some sides parallel to the axis along
which the pieces fit together. When the print with pin-tolerance 0.5
wiggled, in a fit of pique I set pin-tolerance to 0.01. Then some of
these sides of the rotated cubes didn't fit inside each other - think of
trying to stack a perfectly cubical freezer box inside another of the
same size - and the remaining surfaces allowed the pieces to slide
against each other. I'm afraid none of these words make sense until
after you are holding the pieces of plastic; it certainly didn't to me
before it was printed.

Anyway the replacement is a hexagonal cone. This should make sliding
impossible, and make all mating surfaces form an angle with the
fitting-together axis (like a freezer box with angled sides). At the
expense, perhaps, of yet more polygons and CSG items.