make half-circle connectors to connect sides
The x-pins and x-holes (and other x-*) functions generated a D-shaped
connector, centered on the X-axis, with the curve of the D pointing out
the positive Y-axis. They are now generalized and the xu-pins and
xu-holes generate a U-shaped connector, centered on the X-axis, with the
curve of the U pointing down the negative Z-axis. This should hopefully
be useful for connecting the frame of the keyboard to the marshmallowy
sides.