on the precipice of abolishing hollow pins
Hollow pins were supposed to save plastic (and print time) where it
wasn't needed: at the ends of marshmallowy sides pieces. But now I'm
trying to make sides connectors, which will connect the marshmallowy
sides pieces to the frame of the keyboard. To do this, I need to hull a
piece of the side to the hole block. If the hole block has a
half-cylindrical surface at its rear, that's quite easy. If the hole
block is just a wafer, hulling it with something will fill in the holes.
Maybe I won't get rid of hollow pins and wafer holes but they're going
to be redefined completely, and this commit is a savegame before that
happens.