finish sides connectors. eat all RAM and CPU.
This is the commit where I figured out that every time I talk about the
marshmallowy-sides, it adds 5MB to the size of my OpenSCAD output. (The
output from the original adereth/dactyl-keyboard was 500K or so.) So, to
say, "this sides connector is the hull of the intersection of a cone
with the marshmallowy sides, with a small piece of the connector," to
stick the connector to the sides, takes a long time. To talk about all
seven or eight of the sides connectors is prohibitive.