Move key-place out of Clojure into OpenSCAD.
I've manually translated key-place to OpenSCAD; the result, KeyPlace, is
in key-place.scad. I've added to scad-clj a facility for producing a
call to a module including parameters and a block; this facility is
required for this revision to work. In .../dactyl-keyboard, make a
directory called checkouts. In that directory, clone my scad-clj from
https://github.com/jaredjennings/scad-clj, and checkout the
add-module-calls branch. You'll probably need to lein install, and you
may or may not need to remove some cached stuff; I found and removed
this by trial and error.

The end result is that the SCAD code goes from 53MB to 3MB. I think that
if module definitions could be emitted, then (1) key-place.scad could be
generated on the fly; and (2) the key-placed-ribbon could be defined as
a module and save even more space. It'd be nice to get back down to the
neighborhood of the original 500K (!) code.