use pairwise hull rather than minkowski for sides
Now that the sides are built on a shape constructed by hulling together
a bunch of key-placed primitives, there is no reason to make a ribbon
and then minkowski a sphere along it; instead of hulling together
web-posts and then doing something expensive with them, we could just
hull together spheres instead. That's what this change does.

It doesn't end up all that much faster, nor does it make smaller SCAD
code, but the resultant shape looks simpler and nicer.