use subsets of big marshmallow to attach connectors
Previously, the half-circle connectors were hulled to intersections of
the entire big marshmallow sides object with a sliver. This required
mentioning the entire big marshmallow sides object multiple times, which
made the size of the OpenSCAD output balloon, and even after a module
was defined for the object, it still made OpenSCAD's CPU and memory
usage too high.

With this change, we create the part of the marshmallowy sides near to
each connector, and use that for intersection instead, only mentioning
the entire marshmallowy sides object when unioning the half-circle
connectors onto it.