Split bottom cover into pieces (closes #8)
It's nifty how I did this, and this approach came after a lengthy detour
into rhombic dodecahedra, and the honeycomb constructed from them,
turned out to be a dead end. The hexagonal prisms offered here don't cut
sharp angles into the pieces, and the zigzag added over the Z-axis,
combined with the curve of the cover, should offer some keying so that
the pieces aren't quite free to slide along the Z-axis when being glued.
I wanted a more aggressive zig-zag, to be certain that squeezing the
pieces together along the X-axis wouldn't cause them to move shearwise;
but the more aggressive the zig-zag is, the higher the odds that there
will be little detached bits, smaller than 10mm, disembodied in space,
or divoted out and scarcely worthy of gluing. This will certainly be an
area of experimentation. I have no earthly idea how I'm going to clamp
them in any case.