add hacky "feet" (#1)
It's a single foot, really. It might take a long time to print, it's not
going to be good for experimenting with on the bus, if I want to change
the angle a little it will not help me do so, but it's simple, and it
isn't entirely incongruous style-wise. Although it makes me stop
thinking of marshmallows and start thinking of ice cream cones. :)