Add AGPLv3 preambles to pertinent files.
Since I've now written some parts of this in OpenSCAD, and I haven't
really split the manually-written OpenSCAD source code from that which
is generated, the simple statements in the README are no longer
completely unambiguous. These preambles are. I've applied them with
https://github.com/jaredjennings/apply-preambles, so they'll all be the
same, and it should be easy to make your editor fold them up.