shorten output filenames
I sometimes print files via SD card, and the printer's display for
filenames is only so long. With whole words in the filename, like
dactyl-sides-right02.scad, when choosing among multiple files on the SD
card, I see something like "dactyl-sid" or so. So here we shorten names
of output files so they will be distinguishable.