round off edges of frame glue joints
When printing frame pieces, I found that printing them flat leads to
challenges removing the supports ("Remuuve the s'pports!"), but
printing the rectangular shapes rotated up in the air leads to trying
to print a sort of pyramid shape bottom-up. It gets knocked over by
the print head nearly every time. So this is me rounding off those
corners.