Revert "replace lace holes with clamp surfaces"
This reverts commit edf6556a395370075b7989946537670b61f3aa23.

Problems with clamp surfaces as implemented: (1) they are not normal to
the surface being split, but to the xy plane. This means the clamping
surface is sloped relative to the clamp, and clamping will not work. (2)
they indiscriminately fill in features on the surface intended for
gluing things onto. The idea was to sand these clamp surfaces off, but
you can't sand it off of a little wavy screw hole connector. (3) there
was too much clamp surface below the thing to be clamped and not enough
above. Easy fix, but why bother in view of the two other problems?