raise beta; widen bezel; tweak for buildability
It was always clear to me that along a column there should be a curve,
because my fingers move in curves. But it was only recently that I
read why there should be a curve across the row, when my fingers
mostly push down, not out toward the sides of my hand: it is a
compromise to place keys closer together, made possible because
keycaps are narrower at the top. Accordingly in this changeset we
raise beta somewhat.

Test prints of the corner between the thumb and bottom row showed that
the keycap ran into the plastic that joins the thumb to the main
frame. So I widened the bezel. There was one place where that variable
should have been used and wasn't, which is fixed; and there are kluges
in key-place, set to minimize the size of the keyboard while avoiding
keycap collisions, which had to be re-kluged upon this bezel change.

After all this, the keyboard is again more different from the last
revision than I planned, but it should avoid keycap collisions with
the frame, keycap collisions with other keycaps, and be
printable (well, the frame anyway; the whole bottom, sides, screw
holes, and everything else are still completely out of control).